Find free and leading variables.
Also, allow for not changing column order; this doesn't change
solution sets but it does change free variables so our answers
don't match the text.

It's using an "order" hack on the first row where before it was
assuming variables came in alphabetical order.

Maybe try to infer it as long as there is one row that has
all the variables, or if there is another way to get ordering
through a chain of rows that express the full order.